refactor: LectureList 분리
- LectureList 내부에 반응변수가 존재하다보니 필요없는 부분이 리렌더링 되는 문제가 존재
- 해당 문제를 해소하기 위해 View는 재사용하고 Container를 추가
- 이전에 비해 개선되었으나 여전히 렌더링 속도가 좀 느린편
